Code from pangenome pipeline work pre-exodus.
Adds a --multiplicity to limit the number profile combinations
reported.  Previously, all combinations of genomes were reported and
the larger pan genome runs produced a large number of data points,
many of which are redundant for purposes of generating pan genome
plots.  A sampling method was suggested by the Novartis group and
implemented by bwhitty.  This code reflects that effort but has not
been extensively tested.  I need to take a look in more detail and
confirm the sampling is indeed doing something reasonable.